requirements.txt
streamlit>=1.28.0
snowflake-connector-python>=3.5.0
cryptography>=41.0.0
requests>=2.31.0